Initial submission of Registry Monitor Form class
This code add the functionality of running the program in the system tray and pops up a windows notification when changes are detected. It also creates a GUI that lists all changes and allows one or multiple changes to be selected and labeled.
In the Form1_FormClosing event we have added a check that when the user closes the form it will not exit the application but it will hide the form instead.
We added a CellClick event in the dataGridView1 to get the selected row and set it in the Textbox.
We added a button click event btnLabel_Click to update the selected row with the label value entered in the Textbox.